Make type hints cache thread-safe
When the same function is type-checked in multiple threads, a race
condition exists between the full initialization of that function's type
hints in one or more threads and iterating over those hints in one or
more other threads.

This can result in `RuntimeError: OrderedDict mutated during iteration`
in functions that iterate `_CallMemo.type_hints` items.

This fix is to only add the fully initialized type hints dict to the
global type hints cache rather than put an empty dict in the cache and
then subsequently populate that dict.