Use get_protocol_members in protocol checking
This changes `check_protocol()` to make use of `get_protocol_members`
from typing-extensions. This allows removing an existing hard-coded
exclusion list for attributes existing on Protocol, but also handles the
cases `__orig_bases__` and `__weakref__` that was breaking when checking
intersecting protocols (a subclass of two or more protocols).

This has the effect of turning some false positives into true negatives,
but it also leaves some false negatives. To make that clear, xfail test
cases are added for the resulting false negatives.